added version handler
